Refactor and clean up code base
