+++ src/components/CardList/CardList.js
@@ -3,20 +3,19 @@ import { Link } from 'react-router-dom';
 import Card from '../Card/Card';
 
 const CardList = ({ fetchedShows }) => (
-	<div className="row">
-		<div className="col s12">
-			{console.log('showsings', fetchedShows)}
-			{fetchedShows &&
-				fetchedShows.map((show) => (
-					<Link
-						key={show.show.id}
-						to={`/show/${show.show.id}`}
-						show={show.show}>
-						<Card key={show.show.id} show={show.show} />
-					</Link>
-				))}
-		</div>
-	</div>
+  <div className="row">
+    <div className="col s12">
+      {fetchedShows && fetchedShows.map(show => (
+        <Link
+          key={show.show.id}
+          to={`/show/${show.show.id}`}
+          show={show.show}
+        >
+          <Card key={show.show.id} show={show.show} />
+        </Link>
+      ))}
+    </div>
+  </div>
 );
 
 export default CardList;
