--- src/components/Notifications/Notifications.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Notifications = () => (
-  <div>
-    <p>Notifications</p>
-  </div>
-);
-
-export default Notifications;
